Changed int on percentage to float

+++ app/graphql/types/achievement_type.rb
@@ -1,6 +1,6 @@
 class Types::AchievementType < GraphQL::Schema::Object
   field :target_achievement, Int, null: true
   field :current_achievement, Int, null: true
-  field :percentage, Int, null: true
+  field :percentage, Float, null: true
   field :last_month_achievement, Int, null: true
 end
